Improved and centralized the way that the packets are observed
Handling and observing the channel directly everywhere was causing that multiple observers of IPacket Receiver to overlap and cause unexpected behaviors.
Now, a Packet Listener replaced the old channel adapters. This removed the need of having a forced Protocol Channel. Also, the packets and errors in all the layers are centralized in the Observable Packets property of the listener, which can be consumed everywhere without risk.