Don't use git cache for command line builds
Since fetching tags doesn't change the gitinfo inputs, the cache is
used and you may end up with the wrong version #. For IDE builds,
this is not a problem, but for command line builds, you want to be
as precise as possible with versioning, and the GitInfo.Cache file
adds very minimal perf. improvements that were intended more for
dev builds than for CI/cmdline builds anyway.