Fixed Integration Tests
Given that the Server instance is a class instance, it's used by all the tests in the class and if the tests run in parallel then some asserts could fail if we are considering that the connected clients of the server are only the clients that the current test connected.
To avoid this we should always query the Server.ActiveClients property (and not the Server.ActiveConnections) and filtering by the client ids connected in the test.