Ensure the TcpChannel.IsConnected property is always up to date
Sometimes the System.Net.Sockets.TcpClient.Connected returns false positives because as per the documentation, it can be "true if the Client socket was connected to a remote resource as of the most recent operation".
For that reason, if the network stream completes and the observable sequence of the TcpChannel also completes, the value of the property might not be accurate.

To fix this, I'm adding a new "completed" variable to track when the network stream has ended, so it can also be considered as part of the logic to determine if the channel is connected